requirements.txt
tqdm==4.65.0
Pillow
opencv-python
imageio
matplotlib
kornia
natsort
pyyaml
plotly
lpips
open3d==0.16.0
torchmetrics
cyclonedds
pytorch-msssim
jupyter
trimesh
git+https://github.com/JonathonLuiten/diff-gaussian-rasterization-w-depth/